[ISSUE #9651]♻️Extract Tokio client API facades
